services/embedding-dedupe/requirements.txt
fastapi
uvicorn
structlog
pydantic
pydantic-settings
kafka-python
psycopg2-binary
pgvector
sentence-transformers
torch
huggingface_hub
python-dotenv
